Make use of TrimPrefix for binary headers
Currently the Header processing in UnmarshalBinary makes use of TrimLeft which will have undesirable effects when a character in the cutset immediately follows the hyphen.  For example a header value of `CE-CloudEventsVersion` would be converted to `loudEventsVersion`.

This change amends TestHTTPMarshallerFromRequestBinaryBase64Success to add a header to exercise the identifed issue and also changes TrimLeft to TrimPrefix within http.go.

Signed-off-by: Richard Gee <richard@technologee.co.uk>