Don't open and close HTTP connections too often (#110)
* Don't open and close HTTP connections too often

This adds a test to verify we aren't opening and closing HTTP
connections too often when sending events to a single host. The test
will fail in CI (it fails locally) until I push a subsequent change to
actually fix this issue.

The test tracks the count of opened HTTP connections and fails if it's
an amount more than double the expected concurrency. In practice, when
this is fixed properly, the number of opened connections should be
close to if not exactly the expected concurrency. In the case of this
test, that would be 200 new connections. As the code stands today,
when I run this on my laptop, we actually opening over 3000 new
connections in just the 1 second this test runs.

Signed-off-by: Ben Browning <bbrownin@redhat.com>

* Don't forcibly close HTTP connections after every request

Signed-off-by: Ben Browning <bbrownin@redhat.com>

* Allow users to override http.Request.Close value

If a user explicitly needs to close the connection after sending every
event, they can do so similar to how they can override the request
method and headers - by setting the req.Close value in a http.Request
passed and setting that as `transport.Req`.

Signed-off-by: Ben Browning <bbrownin@redhat.com>