Bindings optimization to support binary -> binary + some refactoring (#270)
* [WIP] Binary message transfer without intermediate Event.

Non-functioning code! This is an API proposal for #262

NOTE: the proposal is incompatible with current API (adds a method to Message)

If it needs to be made compatible, the Binary() method can be moved to a new
interface (e.g. BinaryMessage) and the user can cast.

Signed-off-by: Alan Conway <aconway@redhat.com>

* Reorganized bindings interfaces to support binary -> binary encoding
Added Transcoders, to transcode message between sender and receiver
Reimplemented http & amqp binding with new interfaces

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* Renamed stuff
Added translate function, to abstract message reading process

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* Reverted go.sum

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* Formatted translate.go

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

Co-authored-by: Alan Conway <aconway@redhat.com>