Binding multicast implementation (#292)
* Fix scope warnings from golangci-lint

Signed-off-by: Alan Conway <aconway@redhat.com>

* Added UnknownFormat to allow unknown formats to be forwarded.

Signed-off-by: Alan Conway <aconway@redhat.com>

* Fix scope warnings from golangci-lint

Signed-off-by: Alan Conway <aconway@redhat.com>

* Restore StructMessage, complete EventMessage implementation.

StructMessage and EventMessage are intended to be full implementations of the
Message interface that are independent of any transport and can be used to hold
a copy of a message in memory independently of the life of an incoming transport
message (e.g. for multicast, queuing etc.) CopyMessage copies an incoming
Message, the copy can be read many times and is independent of the original
message lifecycle (e.g. you might Finish() the original before you're done
multicasting the copy depending on policy)

They are a lot like your MockXMessages but they are not mocks, they're complete
Message implementations. Possibly there's some code overlap that can be cleaned up.

Please review the code carefully. I like what you've done to reduce copying but
there was a lot to take in, so this is a bit rushed.

One question: I intended for 2 types of message representation - Binary and
Structured. Message.Event() and EventMessage were really just implementations of
Binary encoding - not meant to be a third representation. I think you could drop
Message.Event() now that you have ToEvent(Message) so bindings only have to
implement the methods Structured() and Binary().

Signed-off-by: Alan Conway <aconway@redhat.com>

* Added WithAcksBeforeFinish

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* Reverted uint32 to int32
Test integration CopyMessage + WithAcksBeforeFinish

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* Fixed bunch of things
Added multi senders

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* General refactoring:
* CopyMessage moved in buffering submodule
* CopyMessage splitted to CopyMessage and BufferMessage (one binds the lifecycle, the other no)
* CopyMessage & BufferMessage implements buffer pooling to decrease memory allocations
* EventMessage, ToEvent & Transport adapter moved in event submodule

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* Added benchmark for BufferMessage

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

* Removed context from binary buffered message

Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

Co-authored-by: Alan Conway <aconway@redhat.com>