Add protobuf datacodec (#688)
* Add protobuf datacodec

This adds the ability to send and receive protobuf encoded data within
the envelope.

Signed-off-by: Kevin Conway <kevinconway@invisionapp.com>

* Move pb.Any handling to protobuf format

This refactors the original protobuf datacodec so that it is generally
useful as a codec within other formats. Previously, the coded wrapped
and unwrapped every value in an Any container because this is required
when the protobuf codec and format are used together. Now, the format
handles wrapping the protobuf encoded binary data in the Any type.

Signed-off-by: Kevin Conway <kevinconway@invisionapp.com>

* Add init to register protobuf codec

Signed-off-by: Kevin Conway <kevinconway@invisionapp.com>