Combine request context with parent receiver context
PR #718 changed the handler to always receive the request context in
order to support middleware that alters the context.  This however means
we lose any context values (loggers, etc) that are set on the outer
context.

This change introduces a delegating context that will prioritize the
child (request) context for values and cancellation, but fallback to the
parent context for missing values.

Signed-off-by: Ben Moss <benm@vmware.com>