Support ACK when receiving malformed events
My team faced much the same issue as outlined in #757; malformed events are sent to a Kafka topic
and clients endlessly fail to read the event. While this is hard to induce when the sender uses
the Go CloudEvents SDK, there are a good amount of Python clients across our services which
unfortunately lack validation that might prevent this.

I've elected to make this behaviour configurable via client options, as suggested in #757. This
would be appropriate to use when no `protocol.Responder` implementation is available, as is the
case with the `kafka_sarama` module. I explored wrapping the existing `protocol.Receiver`
implementation to allow it to behave like `protocol.Responder`, but that ended up being a lot of
code compared to the light touch that could be applied here.

Signed-off-by: Matthew Dowdell <matthew.dowdell@hpe.com>