Support content types following structured syntax suffixes
This enables the encoding/decoding of custom content types which
use the `+json` or `+xml` suffix to indicate that their encoding
is JSON or XML respectively. Users can also add their own suffixes
to the registry in a similar fashion to how `datacodec.AddDecoder()`
and `dataodec.AddEncoder()` work, but via the `AddStructuredSuffixDecoder`
and `AddStructuredSuffixEncoder` functions.

Signed-off-by: dan-j <5727701+dan-j@users.noreply.github.com>
Signed-off-by: Noah Kreiger <noahkreiger@gmail.com>